refactor(exercises-pager): Split hook, fix hardcoded text, comply with guidelines
Extract navigation logic into useExercisesPager hook to bring component
under 150-line limit. Replace hardcoded "Back to Chapter" with translated
backToChapter key via new BackToChapter client component.
