+++ src/app/(frontend)/courses/[courseSlug]/chapters/[chapterSlug]/lessons/[lessonSlug]/_components/ExercisesPager/index.tsx
@@ -1,6 +1,5 @@
 'use client'
 
-import { useState } from 'react'
 import type { Exercise } from '@/payload-types'
 import { Button } from '@/ui/web/components/button'
 import { SystemLink } from '@/infra/loading/components/SystemLink'
@@ -9,6 +8,7 @@ import { Card, CardHeader, CardTitle, CardDescription, CardContent } from '@/ui/
 import { ChevronLeft, ChevronRight } from 'lucide-react'
 import { useTranslations } from '@/ui/web/providers/I18n'
 import type { ExerciseContentData } from '@/ui/web/exerciserenderer/types'
+import { useExercisesPager } from './useExercisesPager'
 
 interface ExercisesPagerProps {
   exercises: Exercise[]
@@ -16,73 +16,13 @@ interface ExercisesPagerProps {
   backUrl: string
 }
 
-type PageType = 'intro' | 'exercise' | 'completed'
-
-interface PageState {
-  type: PageType
-  /** 0 = intro, 1..N = exercise index, N+1 = completed */
-  pageNumber: number
-  /** For exercise pages, the exercise being displayed */
-  exerciseIndex?: number
-}
-
 export function ExercisesPager({ exercises, lessonTitle, backUrl }: ExercisesPagerProps) {
   const t = useTranslations('courses')
-
-  const [pageState, setPageState] = useState<PageState>({
-    type: 'intro',
-    pageNumber: 0,
-  })
-
-  const totalExercises = exercises.length
-  const totalPages = totalExercises + 2 // intro + exercises + completed
-
-  const handleNext = () => {
-    const nextPage = pageState.pageNumber + 1
-
-    if (nextPage === totalPages - 1) {
-      // Moving to completed page
-      setPageState({ type: 'completed', pageNumber: nextPage })
-    } else if (nextPage > 0 && nextPage < totalPages - 1) {
-      // Moving to an exercise page
-      setPageState({
-        type: 'exercise',
-        pageNumber: nextPage,
-        exerciseIndex: nextPage - 1,
-      })
-    }
-  }
-
-  const handlePrev = () => {
-    const prevPage = pageState.pageNumber - 1
-
-    if (prevPage === 0) {
-      // Moving back to intro
-      setPageState({ type: 'intro', pageNumber: 0 })
-    } else if (prevPage > 0 && prevPage < totalPages - 1) {
-      // Moving to a previous exercise
-      setPageState({
-        type: 'exercise',
-        pageNumber: prevPage,
-        exerciseIndex: prevPage - 1,
-      })
-    }
-  }
-
-  const handleStart = () => {
-    setPageState({
-      type: 'exercise',
-      pageNumber: 1,
-      exerciseIndex: 0,
-    })
-  }
-
-  const canGoNext = pageState.pageNumber < totalPages - 1
-  const canGoPrev = pageState.pageNumber > 0
+  const { pageState, canGoNext, canGoPrev, handleNext, handlePrev, handleStart } =
+    useExercisesPager(exercises.length)
 
   return (
     <div className="w-full h-full flex flex-col">
-      {/* Main Content Area */}
       <div className="flex-1 overflow-y-auto p-4 md:p-8">
         <div className="max-w-4xl mx-auto">
           {/* Intro Page */}
@@ -91,7 +31,7 @@ export function ExercisesPager({ exercises, lessonTitle, backUrl }: ExercisesPag
               <CardHeader className="text-center pb-4">
                 <CardTitle className="text-3xl md:text-4xl font-bold mb-4">{lessonTitle}</CardTitle>
                 <CardDescription className="text-base md:text-lg">
-                  {t('exercisesPagerIntroDescriptionPart1')} {totalExercises}{' '}
+                  {t('exercisesPagerIntroDescriptionPart1')} {exercises.length}{' '}
                   {t('exercisesPagerIntroDescriptionPart2')}
                 </CardDescription>
               </CardHeader>
@@ -107,19 +47,16 @@ export function ExercisesPager({ exercises, lessonTitle, backUrl }: ExercisesPag
           {/* Exercise Page */}
           {pageState.type === 'exercise' && pageState.exerciseIndex !== undefined && (
             <div className="space-y-6">
-              {/* Exercise Header */}
               <div className="flex items-center justify-between mb-4">
                 <div>
                   <p className="text-sm text-muted-foreground mb-1">
-                    {t('exercise')} {pageState.exerciseIndex + 1} {t('of')} {totalExercises}
+                    {t('exercise')} {pageState.exerciseIndex + 1} {t('of')} {exercises.length}
                   </p>
                   <h2 className="text-2xl md:text-3xl font-bold">
                     {exercises[pageState.exerciseIndex]?.title}
                   </h2>
                 </div>
               </div>
-
-              {/* Exercise Content */}
               <ExerciseRenderer
                 content={
                   exercises[pageState.exerciseIndex]?.content as unknown as ExerciseContentData
@@ -151,10 +88,9 @@ export function ExercisesPager({ exercises, lessonTitle, backUrl }: ExercisesPag
         </div>
       </div>
 
-      {/* Navigation Controls - Fixed at bottom */}
+      {/* Navigation Controls */}
       <div className="border-t bg-background/95 backdrop-blur supports-[backdrop-filter]:bg-background/60">
         <div className="max-w-4xl mx-auto px-4 py-4 flex items-center justify-between gap-4">
-          {/* Previous Button */}
           <Button
             onClick={handlePrev}
             disabled={!canGoPrev}
@@ -166,15 +102,13 @@ export function ExercisesPager({ exercises, lessonTitle, backUrl }: ExercisesPag
             {t('exercisesPagerPrev')}
           </Button>
 
-          {/* Page Indicator */}
           <div className="text-sm text-muted-foreground hidden sm:block">
             {pageState.type === 'intro' && t('exercisesPagerIntro')}
             {pageState.type === 'exercise' &&
-              `${t('exercise')} ${pageState.exerciseIndex! + 1}/${totalExercises}`}
+              `${t('exercise')} ${pageState.exerciseIndex! + 1}/${exercises.length}`}
             {pageState.type === 'completed' && t('exercisesPagerCompleted')}
           </div>
 
-          {/* Next Button */}
           <Button onClick={handleNext} disabled={!canGoNext} size="lg" className="min-w-[120px]">
             {t('exercisesPagerNext')}
             <ChevronLeft className="ms-2 h-5 w-5 rtl:rotate-0 ltr:rotate-180" />
