+++ src/i18n/en.json
@@ -108,6 +108,7 @@
     "featuredTitle": "Featured Courses",
     "featuredDescription": "Hand-picked courses to help you on your learning journey.",
     "backToCourses": "Back to Courses",
+    "backToChapter": "Back to Chapter",
     "openCourse": "Open Course",
     "viewChapter": "View Chapter",
     "viewLesson": "View Lesson",
