+++ src/i18n/he.json
@@ -108,6 +108,7 @@
     "featuredTitle": "קורסים מומלצים",
     "featuredDescription": "קורסים שנבחרו בקפידה כדי לעזור לך במסע הלמידה שלך.",
     "backToCourses": "חזרה לקורסים",
+    "backToChapter": "חזרה לפרק",
     "openCourse": "פתח קורס",
     "viewChapter": "צפה בפרק",
     "viewLesson": "צפה בשיעור",
